Fix bug when GameMaster used LoadScene on ToolsUI instead of LoadSceneManager

+++ Assets/Scripts/Managers/Timer.cs
@@ -48,6 +48,6 @@ public class Timer : MonoBehaviour {
 
     string ConvertTimeLeft()
     {
-        return (int)(TimeLeft / 60) + ":" + ((int)(TimeLeft % 60)).ToString("00");
+        return (int)(TimeLeft / 60) + ":" + (TimeLeft % 60).ToString("00");
     }
 }
